Re-organise stategies into own files
