+++ aquadoggo/src/proptests/mod.rs
@@ -1,3 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+pub mod schema_strategies;
+pub mod document_strategies;
+pub mod utils;
 mod tests;
